frontend/requirements.txt
streamlit
python-dotenv
requests
sseclient-py
